Supply Chain Optimization System/requirements.txt
pandas
numpy
psycopg2-binary
scikit-learn
joblib
mlflow
kafka-python
streamlit
plotly
requests